extract notebook to commutable/src/notebook.ts
